requirements.txt
pandas
numpy
matplotlib
librosa==0.7.2
scikit-learn==0.22.2.post1
torch==1.7.1
math
pickle
seaborn
umap-learn
